+++ src/__tests__/bootstrap.test.js
@@ -7,7 +7,24 @@ import bootstrap from '../bootstrap';
 jest.mock('../utils/shell');
 jest.mock('../utils/writeSpecs');
 
-const PACKAGE_NAMES = ['oao', 'oao-b', 'oao-c', 'oao-d', 'oao-priv'];
+const PACKAGE_NAMES_1 = ['oao', 'oao-b', 'oao-c', 'oao-d', 'oao-priv'];
+const PACKAGE_NAMES_2 = ['oao', 'oao-b', 'oao-c'];
+
+const readOriginalSpecs = (base, names) => {
+  const originalSpecs = {};
+  names.forEach((name) => {
+    originalSpecs[name] = require(path.join(base, `${name}/package.json`));
+  });
+  return originalSpecs;
+};
+
+const spyFinalSpec = (spy, pkgName) => {
+  let finalSpec;
+  spy.mock.calls.forEach(([, specs]) => {
+    if (specs.name === pkgName) finalSpec = specs;
+  });
+  return finalSpec;
+};
 
 describe('BOOTSTRAP command', () => {
   beforeEach(() => { jest.resetAllMocks(); });
@@ -15,34 +32,45 @@ describe('BOOTSTRAP command', () => {
   it('does not modify any package.json', async () => {
     const writeSpecs = require('../utils/writeSpecs').default;
     const base = path.join(process.cwd(), 'test/fixtures/packages');
-
-    // Copy original specs
-    const originalSpecs = {};
-    PACKAGE_NAMES.forEach((name) => {
-      originalSpecs[name] = require(path.join(base, `${name}/package.json`));
-    });
-
-    // Run command
+    const originalSpecs = readOriginalSpecs(base, PACKAGE_NAMES_1);
     await bootstrap({ src: 'test/fixtures/packages/*' });
-
     // Process call arguments; we keep the last time a spec is stored
     // and compare it with the original one
-    const finalSpecs = {};
     writeSpecs.mock.calls.forEach(([specPath, specs]) => {
       const { name } = specs;
-      finalSpecs[name] = specs;
       expect(specPath.split(path.sep)).toContain(name);
       expect(specPath.split(path.sep)).toContain('package.json');
     });
-    expect(Object.keys(finalSpecs)).toHaveLength(PACKAGE_NAMES.length);
-    PACKAGE_NAMES.forEach((name) => {
-      expect(finalSpecs[name]).toEqual(originalSpecs[name]);
+    PACKAGE_NAMES_1.forEach((name) => {
+      expect(spyFinalSpec(writeSpecs, name)).toEqual(originalSpecs[name]);
+    });
+  });
+
+  it('does not modify any package.json with custom links', async () => {
+    const writeSpecs = require('../utils/writeSpecs').default;
+    const base = path.join(process.cwd(), 'test/fixtures/packagesCustomLinks');
+    const originalSpecs = readOriginalSpecs(base, PACKAGE_NAMES_2);
+    await bootstrap({
+      src: 'test/fixtures/packagesCustomLinks/*',
+      link: 'ext-.*',
+    });
+    PACKAGE_NAMES_2.forEach((name) => {
+      expect(spyFinalSpec(writeSpecs, name)).toEqual(originalSpecs[name]);
     });
   });
 
   it('executes the correct `yarn link`s and `yarn install`s', async () => {
     const helpers = require('../utils/shell');
     await bootstrap({ src: 'test/fixtures/packages/*' });
+    expect(helpers.exec.mock.calls).toMatchSnapshot();
+  });
+
+  it('executes the correct `yarn link`s and `yarn install`s with custom links', async () => {
+    const helpers = require('../utils/shell');
+    await bootstrap({
+      src: 'test/fixtures/packagesCustomLinks/*',
+      link: 'ext-.*',
+    });
     expect(helpers.exec.mock.calls).toMatchSnapshot();
   });