+++ src/all.js
@@ -3,7 +3,7 @@
 import listPaths from './utils/listPaths';
 import { exec } from './utils/shell';
 
-type Options = {| src: string |};
+type Options = { src: string };
 
 const run = async (cmd: string, { src: srcPatterns }: Options) => {
   const pkgPaths = await listPaths(srcPatterns);